requirements.txt
websockets==12.0
requests==2.32.3
python-dotenv==1.0.1
pycryptodome==3.20.0
pytest==7.4.3
pytest-asyncio==0.21.1